streamlit_app/requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.0
matplotlib==3.8.2
jupyter==1.0.0
ipykernel==6.29.0
pandas==2.2.0
